payment done

+++ src/main/java/ru/snptech/businessbanyabot/integrations/bank/dto/request/qr/RegisterPaymentQrCodeData.java
@@ -9,6 +9,6 @@ public record RegisterPaymentQrCodeData(
     String qrcType,
     QrImage imageParams,
     String sourceName,
-    Integer ttl
+    Long ttl
 ) {
 }
